% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albacore}
\alias{albacore}
\title{Albacore data}
\format{A data frame with 23 rows and 4 variables:
\describe{
  \item{Year}{Year of catch}
  \item{C}{Catches (1000 t)}
  \item{E}{Effort (10^8 hooks)}
  \item{I}{CPUE (kg/100 hooks)}
  \item{P}{P = B/K}
}}
\usage{
albacore
}
\description{
South Atlantic albacore data from Millar & Meyer (2001).
}
\keyword{datasets}
