% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake}
\alias{hake}
\title{Namibian hake data}
\format{A data frame with 24 rows and 4 variables:
\describe{
  \item{year}{Year of catch}
  \item{catch}{Catches (1000 t)}
  \item{effort}{Effort}
  \item{cpue}{CPUE}
}}
\usage{
hake
}
\description{
Northern Namibian hake data.
}
\keyword{datasets}
