% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dmfs.R
\name{plot.dmfs}
\alias{plot.dmfs}
\title{Plot simulated data}
\usage{
\method{plot}{dmfs}(x, qscale, ...)
}
\arguments{
\item{x}{An object of class \code{dmfs}.}

\item{qscale}{The parameter \eqn{q} (catchability coefficient). You
may use the same value used for the simulation, so that biomass
and CPUE values will be in the same scale (just for
visualization).}

\item{...}{Not used.}
}
\value{
Plots of simulated dynamics.
}
\description{
Plot simulated data for objects of class \code{dmfs}.
}
\details{
Use this function to visualize data simulated with
    \code{simdata()}.
}
\examples{
set.seed(1)
sim <- simdata(q = 0.01)
plot(sim, qscale = 0.01)
}
\author{
Fernando Mayer
}
