% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.gen.R
\name{schaefer.gen}
\alias{schaefer.gen}
\title{Generate biomass values form Schaefer's model}
\usage{
schaefer.gen(Binit, r, K, C)
}
\arguments{
\item{Binit}{Initial biomass}

\item{r}{Intrinsic growth rate}

\item{K}{Population's carrying capacity}

\item{C}{Observed catch data}
}
\value{
A vector of biomass values with the same length as \code{C}
}
\description{
Generate biomass values based on Schaefer model
}
\details{
Based in a set of parameters, generate a sequence of biomass values
    based on the Schaefer model.
}
\examples{
B <- schaefer.gen(Binit = 2500, r = 0.8, K = 3000,
C = runif(20, 70, 900))
plot(B, type = "l")
}
\author{
Fernando Mayer
}
