% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.procobs.R
\name{schaefer.procobs}
\alias{schaefer.procobs}
\title{Negative log-likelihood function for the Schaefer model}
\usage{
schaefer.procobs(par, B, I, C)
}
\arguments{
\item{par}{A (named) vector with starting values of parameters:
\code{r}, \code{K}, \code{q}, \code{sigmaproc} (process standard
deviation), \code{sigmaobs} (observation standard deviation).}

\item{B}{A vector of biomass estimates calculated based on the
initial parameters.}

\item{I}{A vector of observed CPUE values.}

\item{C}{A vector of observed catch values.}
}
\value{
The negative log-likelihood value for a given set of
    parameters and data.
}
\description{
Negative log-likelihood function for the Schaefer model with both
process and observation errors.
}
\details{
This function calculates the negative log-likelihood for the
    Schaefer model, considering multiplicative log-normal errors for
    the process and observation equation.

The function is suited for use in \code{optim()}, \code{nlminb()},
    or \code{bbmle::mle2()} (with minor modifications). The
    minimization will be made in the sum of both likelihhods.
}
\author{
Fernando Mayer
}
