% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple}
\alias{simple}
\title{A simple fake dataset}
\format{A data frame with 18 rows and 4 variables:
\describe{
  \item{Year}{Year of catch}
  \item{Catch}{Catches in t}
  \item{Eff}{Effort in fisherman/day}
  \item{CPUE}{CPUE in t/(fisherman/day)}
}}
\usage{
simple
}
\description{
A random catch and effort dataset just for testing functions.
}
\keyword{datasets}
