% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dmfs.R
\name{plot.dmfs}
\alias{plot.dmfs}
\title{Plot simulated data}
\usage{
\method{plot}{dmfs}(x, qscale, ...)
}
\arguments{
\item{x}{An object of class \code{dmfs}.}

\item{qscale}{The parameter \eqn{q} (catchability coefficient), used
to scale CPUE values so that they are at the same scale as
biomass values (this is just for visualization purposes). If
nothing is specified, then by default it will be used the same
value for \code{q} used in the \code{simdata()} function
(recommended).}

\item{...}{Not used.}
}
\value{
Plots of simulated dynamics.
}
\description{
Plot simulated data for objects of class \code{dmfs}.
}
\details{
Use this function to visualize data simulated with
    \code{simdata()}.
}
\examples{
set.seed(1)
sim <- simdata()
plot(sim)
}
\author{
Fernando Mayer
}
