% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.gen.R
\name{schaefer.gen}
\alias{schaefer.gen}
\title{Generate biomass values form Schaefer's model}
\usage{
schaefer.gen(r, K, C, sd.proc)
}
\arguments{
\item{r}{Intrinsic growth rate}

\item{K}{Population's carrying capacity. This value will also be
used as starting value for biomass \eqn{B_1} (plus error).}

\item{C}{Observed catch data}

\item{sd.proc}{Process standard deviation.}
}
\value{
A vector of biomass values with the same length as \code{C}
}
\description{
Generate biomass values based on Schaefer model
}
\details{
Based in a set of parameters, generate a sequence of biomass values
    based on the Schaefer model.
}
\examples{
set.seed(1)
B <- schaefer.gen(r = 0.8, K = 3000,
                  C = runif(20, 70, 900), sd.proc = 0.05)
plot(B, type = "l")
}
\author{
Fernando Mayer
}
